# docs/swagger.yaml
basePath: /
definitions:
...
  config.EffectiveResponse:
    properties:
      effectiveConfig:
        additionalProperties:
          $ref: '#/definitions/model.EffectiveConfigItem'
        type: object
    type: object
...
  httpx.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  model.EffectiveConfigItem:
    properties:
      key:
        type: string
      scope:
        type: string
      value:
        type: string
      valueType:
        type: string
    type: object
...
info:
  contact:
    name: etcdisc
  description: etcdisc phase-1 control plane API for registry, discovery, config,
    A2A, and admin operations.
  title: etcdisc HTTP API
  version: "1.0"
paths:
...
  /v1/config/effective:
    get:
      description: Return effective config after global, namespace, and service overlays
        are applied.
      parameters:
      - description: namespace
        in: query
        name: namespace
        required: true
        type: string
      - description: service
        in: query
        name: service
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.EffectiveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get effective config
      tags:
      - config
  /v1/config/watch:
    get:
      description: Open an SSE stream for config change events.
      parameters:
      - description: namespace
        in: query
        name: namespace
        required: true
        type: string
      - description: service
        in: query
        name: service
        required: true
        type: string
      - description: start revision
        in: query
        name: revision
        type: integer
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
      summary: Watch config changes
      tags:
      - config
...
schemes:
- http
swagger: "2.0"